Make experimental optics a checkbox
